expense-tracker/requirements.txt
sqlalchemy
matplotlib
tkcalendar
seaborn
pandas
numpy
Pillow